+++ src/pages/index.tsx
@@ -2,8 +2,8 @@ import { JoinDiscordServer } from '../components/joinDiscordServer';
 import { Head } from '../components/head';
 import { Blogs } from '../components/blogs';
 import { Faq } from '../components/faq';
-import { Contact } from '../components/contact/contact';
 import { Descriptions } from '../components/descriptions';
+import { Contact } from '../components/contact/contact';
 
 export default async function HomePage() {
   return (
